Implement user registration and authentication routes

+++ app/resources/note.py
@@ -1,6 +1,7 @@
 from flask_smorest import Blueprint
 from flask.views import MethodView
 from flask import request
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from app import db
 from app.models import Note
@@ -19,21 +20,39 @@ class NoteResource(MethodView):
         db.session.commit()
         return note_schema.dump(note), 201
 
+    @jwt_required()  # Protect this route with JWT
     def get(self, note_id):
+        current_user = get_jwt_identity()
         note = Note.query.get_or_404(note_id)
+
+        if note.user_id != current_user:
+            return '', 403  # Forbidden
+
         note_schema = NoteSchema()
         return note_schema.dump(note), 200
 
+    @jwt_required()  # Protect this route with JWT
     def put(self, note_id):
+        current_user = get_jwt_identity()
         note = Note.query.get_or_404(note_id)
         data = request.get_json()
+
+        if note.user_id != current_user:
+            return '', 403  # Forbidden
+
         note_schema = NoteSchema()
         note = note_schema.load(data, instance=note)
         db.session.commit()
         return note_schema.dump(note), 200
 
+    @jwt_required()  # Protect this route with JWT
     def delete(self, note_id):
+        current_user = get_jwt_identity()
         note = Note.query.get_or_404(note_id)
+
+        if note.user_id != current_user:
+            return '', 403  # Forbidden
+
         db.session.delete(note)
         db.session.commit()
         return '', 204
